+++ BackupRestore/BackupRestore.php
@@ -46,7 +46,7 @@ use PDO;
  * ]);
  *
  * @package BackupRestore
- * @version 0.2.0
+ * @version 0.3.0
  * @license MIT
  */
 class BackupRestore